+++ readium/navigator/src/main/java/org/readium/r2/navigator/R2BasicWebView.kt
@@ -65,8 +65,8 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
         fun onProgressionChanged()
         fun onHighlightActivated(id: String)
         fun onHighlightAnnotationMarkActivated(id: String)
-        fun goForward(animated: Boolean = false, completion: () -> Unit = {}): Boolean
-        fun goBackward(animated: Boolean = false, completion: () -> Unit = {}): Boolean
+        fun goForward(animated: Boolean = false, completion: () -> Unit = {}): Boolean = false
+        fun goBackward(animated: Boolean = false, completion: () -> Unit = {}): Boolean = false
 
         /**
          * Returns the custom [ActionMode.Callback] to be used with the text selection menu.
@@ -75,18 +75,26 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
 
         @InternalReadiumApi
         fun javascriptInterfacesForResource(link: Link): Map<String, Any?> = emptyMap()
-
         @InternalReadiumApi
         fun shouldOverrideUrlLoading(webView: WebView, request: WebResourceRequest): Boolean = false
-
         @InternalReadiumApi
         fun shouldInterceptRequest(webView: WebView, request: WebResourceRequest): WebResourceResponse? = null
-
         @InternalReadiumApi
         fun resourceAtUrl(url: String): Resource? = null
+
+        /**
+         * Requests to load the next resource in the reading order.
+         *
+         * @param jump Indicates whether it's a discontinuous jump from the current locator. Used
+         * for scroll mode.
+         */
+        @InternalReadiumApi
+        fun goToNextResource(jump: Boolean, animated: Boolean): Boolean = false
+        @InternalReadiumApi
+        fun goToPreviousResource(jump: Boolean, animated: Boolean): Boolean = false
     }
 
-    lateinit var listener: Listener
+    var listener: Listener? = null
     internal var preferences: SharedPreferences? = null
     internal var useLegacySettings: Boolean = false
 
@@ -124,7 +132,7 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
             val pageWidth = computeHorizontalScrollExtent()
             val contentWidth = computeHorizontalScrollRange()
 
-            val isRtl = (listener.readingProgression == ReadingProgression.RTL)
+            val isRtl = (listener?.readingProgression == ReadingProgression.RTL)
 
             // For RTL, we need to add the equivalent of one page to the x position, otherwise the
             // progression will be one page off.
@@ -162,7 +170,7 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
 
     override fun onScrollChanged(l: Int, t: Int, oldl: Int, oldt: Int) {
         super.onScrollChanged(l, t, oldl, oldt)
-        listener.onProgressionChanged()
+        listener?.onProgressionChanged()
     }
 
     override fun onAttachedToWindow() {
@@ -180,24 +188,34 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
     @android.webkit.JavascriptInterface
     open fun scrollRight(animated: Boolean = false) {
         uiScope.launch {
+            val listener = listener ?: return@launch
             listener.onScroll()
 
-            fun goRight() {
+            val isRtl = (listener.readingProgression == ReadingProgression.RTL)
+
+            fun goRight(jump: Boolean) {
                 if (listener.readingProgression == ReadingProgression.RTL) {
-                    listener.goBackward(animated = animated)
+                    listener.goBackward(animated = animated) // Legacy
+                    listener.goToPreviousResource(jump = jump, animated = animated)
                 } else {
-                    listener.goForward(animated = animated)
+                    listener.goForward(animated = animated) // Legacy
+                    listener.goToNextResource(jump = jump, animated = animated)
                 }
             }
 
-            if (scrollMode || !this@R2BasicWebView.canScrollHorizontally(1)) {
-                goRight()
-            } else {
-                runJavaScript("readium.scrollRight();") { success ->
-                    if (!success.toBoolean()) {
-                        goRight()
+            when {
+                scrollMode ->
+                    goRight(jump = true)
+
+                !this@R2BasicWebView.canScrollHorizontally(1) ->
+                    goRight(jump = false)
+
+                else ->
+                    runJavaScript("readium.scrollRight();") { success ->
+                        if (!success.toBoolean()) {
+                            goRight(jump = false)
+                        }
                     }
-                }
             }
         }
     }
@@ -205,24 +223,32 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
     @android.webkit.JavascriptInterface
     open fun scrollLeft(animated: Boolean = false) {
         uiScope.launch {
+            val listener = listener ?: return@launch
             listener.onScroll()
 
-            fun goLeft() {
+            fun goLeft(jump: Boolean) {
                 if (listener.readingProgression == ReadingProgression.RTL) {
-                    listener.goForward(animated = animated)
+                    listener.goForward(animated = animated) // legacy
+                    listener.goToNextResource(jump = jump, animated = animated)
                 } else {
-                    listener.goBackward(animated = animated)
+                    listener.goBackward(animated = animated) // legacy
+                    listener.goToPreviousResource(jump = jump, animated = animated)
                 }
             }
 
-            if (scrollMode || !this@R2BasicWebView.canScrollHorizontally(-1)) {
-                goLeft()
-            } else {
-                runJavaScript("readium.scrollLeft();") { success ->
-                    if (!success.toBoolean()) {
-                        goLeft()
+            when {
+                scrollMode ->
+                    goLeft(jump = true)
+
+                !this@R2BasicWebView.canScrollHorizontally(-1) ->
+                    goLeft(jump = false)
+
+                else ->
+                    runJavaScript("readium.scrollLeft();") { success ->
+                        if (!success.toBoolean()) {
+                            goLeft(jump = false)
+                        }
                     }
-                }
             }
         }
     }
@@ -270,7 +296,7 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
                 true
             }
             else ->
-                runBlocking(uiScope.coroutineContext) { listener.onTap(event.point) }
+                runBlocking(uiScope.coroutineContext) { listener?.onTap(event.point) ?: false }
         }
     }
 
@@ -289,7 +315,7 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
             return false
         }
 
-        return listener.onDecorationActivated(id, group, rect, click.point)
+        return listener?.onDecorationActivated(id, group, rect, click.point) ?: false
     }
 
     /** Produced by gestures.js */
@@ -336,7 +362,7 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
 
         val aside = runBlocking {
             tryOrLog {
-                listener.resourceAtUrl(absoluteUrl)
+                listener?.resourceAtUrl(absoluteUrl)
                     ?.use { res ->
                         res.readAsString()
                             .map { Jsoup.parse(it) }
@@ -398,7 +424,7 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
         val event = DragEvent.fromJSON(eventJson)?.takeIf { it.isValid }
             ?: return false
 
-        return runBlocking(uiScope.coroutineContext) { listener.onDragStart(event) }
+        return runBlocking(uiScope.coroutineContext) { listener?.onDragStart(event) ?: false }
     }
 
     @android.webkit.JavascriptInterface
@@ -406,7 +432,7 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
         val event = DragEvent.fromJSON(eventJson)?.takeIf { it.isValid }
             ?: return false
 
-        return runBlocking(uiScope.coroutineContext) { listener.onDragMove(event) }
+        return runBlocking(uiScope.coroutineContext) { listener?.onDragMove(event) ?: false }
     }
 
     @android.webkit.JavascriptInterface
@@ -414,7 +440,7 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
         val event = DragEvent.fromJSON(eventJson)?.takeIf { it.isValid }
             ?: return false
 
-        return runBlocking(uiScope.coroutineContext) { listener.onDragEnd(event) }
+        return runBlocking(uiScope.coroutineContext) { listener?.onDragEnd(event) ?: false }
     }
 
     @android.webkit.JavascriptInterface
@@ -484,14 +510,14 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
     @android.webkit.JavascriptInterface
     fun highlightActivated(id: String) {
         uiScope.launch {
-            listener.onHighlightActivated(id)
+            listener?.onHighlightActivated(id)
         }
     }
 
     @android.webkit.JavascriptInterface
     fun highlightAnnotationMarkActivated(id: String) {
         uiScope.launch {
-            listener.onHighlightAnnotationMarkActivated(id)
+            listener?.onHighlightAnnotationMarkActivated(id)
         }
     }
 
@@ -587,7 +613,7 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
     internal fun shouldOverrideUrlLoading(request: WebResourceRequest): Boolean {
         if (resourceUrl == request.url?.toString()) return false
 
-        return listener.shouldOverrideUrlLoading(this, request)
+        return listener?.shouldOverrideUrlLoading(this, request) ?: false
     }
 
     internal fun shouldInterceptRequest(webView: WebView, request: WebResourceRequest): WebResourceResponse? {
@@ -598,7 +624,7 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
             }
         }
 
-        return listener.shouldInterceptRequest(webView, request)
+        return listener?.shouldInterceptRequest(webView, request)
     }
 
     // Text selection ActionMode overrides
@@ -608,7 +634,7 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
     // used by the web view.
 
     override fun startActionMode(callback: ActionMode.Callback?): ActionMode? {
-        val customCallback = listener.selectionActionModeCallback
+        val customCallback = listener?.selectionActionModeCallback
             ?: return super.startActionMode(callback)
 
         val parent = parent ?: return null
@@ -628,7 +654,7 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
 
     @RequiresApi(Build.VERSION_CODES.M)
     override fun startActionMode(callback: ActionMode.Callback?, type: Int): ActionMode? {
-        val customCallback = listener.selectionActionModeCallback
+        val customCallback = listener?.selectionActionModeCallback
             ?: return super.startActionMode(callback, type)
 
         val parent = parent ?: return null