+++ test-app/src/main/java/org/readium/r2/testapp/Application.kt
@@ -7,7 +7,6 @@
 package org.readium.r2.testapp
 
 import android.content.*
-import android.os.IBinder
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.preferencesDataStore
@@ -23,9 +22,6 @@ import timber.log.Timber
 
 class Application : android.app.Application() {
 
-    val Context.navigatorPreferences: DataStore<Preferences>
-        by preferencesDataStore(name = "navigator-preferences")
-
     lateinit var readium: Readium
         private set
 
@@ -40,26 +36,8 @@ class Application : android.app.Application() {
     private val coroutineScope: CoroutineScope =
         MainScope()
 
-    private val mediaServiceBinder: CompletableDeferred<MediaService.Binder> =
-        CompletableDeferred()
-
-    private val mediaServiceConnection = object : ServiceConnection {
-
-        override fun onServiceConnected(name: ComponentName?, service: IBinder) {
-            Timber.d("MediaService bound.")
-            mediaServiceBinder.complete(service as MediaService.Binder)
-        }
-
-        override fun onServiceDisconnected(name: ComponentName) {
-            Timber.d("MediaService disconnected.")
-            // Should not happen, do nothing.
-        }
-
-        override fun onNullBinding(name: ComponentName) {
-            Timber.d("Failed to bind to MediaService.")
-            // Should not happen, do nothing.
-        }
-    }
+    private val Context.navigatorPreferences: DataStore<Preferences>
+        by preferencesDataStore(name = "navigator-preferences")
 
     override fun onCreate() {
         super.onCreate()
@@ -71,30 +49,27 @@ class Application : android.app.Application() {
         storageDir = computeStorageDir()
 
         /*
-         * Starting media service.
-         */
-
-        // MediaSessionService.onBind requires the intent to have a non-null action.
-        val intent = Intent(MediaService.SERVICE_INTERFACE)
-            .apply { setClass(applicationContext, MediaService::class.java) }
-        startService(intent)
-        bindService(intent, mediaServiceConnection, 0)
-
-        /*
          * Initializing repositories
          */
         bookRepository =
             BookDatabase.getDatabase(this).booksDao()
-                .let { BookRepository(it) }
+                .let { dao ->
+                    BookRepository(
+                        applicationContext,
+                        dao,
+                        storageDir,
+                        readium.lcpService,
+                        readium.streamer
+                    )
+                }
 
         readerRepository =
             coroutineScope.async {
                 ReaderRepository(
                     this@Application,
                     readium,
-                    mediaServiceBinder.await(),
                     bookRepository,
-                    navigatorPreferences
+                    navigatorPreferences,
                 )
             }
     }
@@ -112,6 +87,3 @@ class Application : android.app.Application() {
         )
     }
 }
-
-val Context.resolver: ContentResolver
-    get() = applicationContext.contentResolver